+++ src/models/articleDetail.ts
@@ -1,3 +1,8 @@
+export interface IArticleInfo {
+  newsletter: IArticleDetail;
+  previousNewsletter: IArticleDetail | null;
+  nextNewsletter: IArticleDetail | null;
+}
 export interface IArticleDetail {
   id: number;
   title: string;
